+++ src/ma/eugene/nextbusapi/RouteConfigParser.java
@@ -19,7 +19,7 @@ public class RouteConfigParser extends DefaultHandler {
             String title = attributes.getValue("title");
             String lon = attributes.getValue("lon");
             String lat = attributes.getValue("lat");
-            stops.add(new RouteConfigInfo(title, new Integer(stopId), new Float(lon), new Float(lat)));
+            stops.add(new RouteConfigInfo(title, new Integer(stopId), new Double(lon), new Double(lat)));
         }
     }
 }
